Refactor gclient net implementation
