+++ .github/workflows/run_test.yaml
@@ -28,14 +28,7 @@ jobs:
         pip install -r requirements-dev.txt
     - name: Test with pytest
       run: |
-        mkdir tmp
-        pushd tmp
-        git clone https://github.com/vesoft-inc/nebula-docker-compose.git
-        pushd nebula-docker-compose/
-        cp ../../tests/docker-compose.yaml ./
         docker-compose up -d
         sleep 60
-        popd
-        popd
-        cd tests
         pytest -s -v
+      working-directory: tests
